requirements.txt
discord
python-dotenv
apscheduler
supabase